Create ProSoft Chain
ProSoft Chain is a Delegated Proof-of-Stake, layer-1 blockchain with a vision for a decentralised computation accessible on-chain.

The PROSOFT Chain is a Layer 1 network, providing scalable distributed computing direct to smart contracts or to any bridged blockchain. PROSOFT connects blockchain developers and services to a global pool of computing power. This enables up to 10x more cost-effective computing, greener computing and access to a global network to power your next Metaverse, Dynamic NFT, AI, Machine Learning, DeFi or any computationally demanding dApps and smart contracts.